+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## 2.1.1 under development
 
-- Enh #44: In methods of array definitions add autowiring and improve variadic arguments support (@vjik) 
+- Enh #44: In methods of array definitions add autowiring and improve variadic arguments support (@vjik)
+- Enh #41: Raise minimum PHP version to 8.0 and refactor code (@xepozz, @vjik)
 
 ## 2.1.0 October 25, 2022