+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Enh #46: In definition validator add a check of method name in array definitions (@vjik) 
 - Enh #44: In methods of array definitions add autowiring and improve variadic arguments support (@vjik)
 - Enh #41: Raise minimum PHP version to 8.0 and refactor code (@xepozz, @vjik)
+- Bug #48: Definition validator returns false positive result on empty string (@vjik) 
 - Chg #49: Change result format of `DefinitionStorage::getBuildStack()` method to definition IDs array (@vjik) 
 
 ## 2.1.0 October 25, 2022